feat: update coordinator common function

+++ coordinator/internal/config/config.go
@@ -22,7 +22,7 @@ type RollerManagerConfig struct {
 	RollersPerSession uint8 `json:"rollers_per_session"`
 	// Number of attempts that a session can be retried if previous attempts failed.
 	// Currently we only consider proving timeout as failure here.
-	SessionAttempts uint8 `json:"session_attempts,omitempty"`
+	SessionAttempts int `json:"session_attempts,omitempty"`
 	// Zk verifier config.
 	Verifier *VerifierConfig `json:"verifier,omitempty"`
 	// Proof collection time (in minutes).
